Rename Checker to NetworkStatusChecker, extract to file, enable NWPath status checker
